% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations to run the sampling.}

\item{n}{Number of trials per sample.}

\item{p}{Probability that a single event is successful.}
}
\value{
A numeric vector with the probabilities of having i successes, for all values
of i from 0 to n.
}
\description{
Create a plot and return a table of the probabilities of getting i successes
 out of n trials with probability p of success, for all values of i from 0 to n.
}
\examples{
mybin(100, 10, 0.8)
}
