% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{number of iterations of bootstrapping}

\item{x}{random sample}

\item{fun}{the function describing the statistic to measure}

\item{alpha}{the significance level}

\item{...}{extra parameters for the histogram}
}
\value{
A list containing:
\describe{
  \item{ci}{Numeric vector of length 2 with the lower and upper bounds of the confidence interval.}
  \item{fun}{The function used to compute the statistic (e.g., `"mean"`).}
  \item{x}{The original data sample used for bootstrapping.}
  \item{xstat}{A numeric vector of bootstrap sample statistics, one for each iteration.}
}
}
\description{
Use bootstrapping to find a confidence interval for a statistic using a random sample.
}
\examples{
myboot2(iter=1000, x=runif(20, 1, 10), fun="mean", alpha=0.05)
}
