% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{a function that takes in a single sample and a parameter that describes the log-likelihood function.}

\item{x}{a vector of numbers that are the results of sampling some distribution.}

\item{param}{a vector of numbers that represents the different parameter values you want to try in the grid approximation method}

\item{...}{extra parameters for plotting}
}
\value{
A list with the following components:
\describe{
  \item{i}{The index corresponding to the maximum likelihood (integer).}
  \item{parami}{The parameter value corresponding to the maximum likelihood (numeric).}
  \item{yi}{The maximum log-likelihood value (numeric).}
  \item{slope}{A numeric vector giving slope estimates around the maximum likelihood.}
}
}
\description{
Use the grid approximation method to estimate the parameter that maximizes the log-likelihood function "lfun"
}
\examples{
# Define a simple log-likelihood function
lfun <- function(x, p) {
  dbinom(x, size = 20, prob = p, log = TRUE)
}

# Example data
x <- c(3, 3, 4, 3, 4, 5, 5, 4)

# Find the maximum likelihood
mymaxlik(lfun, x, seq(from = 0, to = 1, length.out = 100))
}
