% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My Normal Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{normal distribution mean}

\item{sigma}{normal distribution standard deviation}

\item{a}{threshold}
}
\value{
list containing these elements
\describe{
  \item{\code{mu}}{The normal distribution mean.}
  \item{\code{sigma}}{The normal distribution standard deviation.}
  \item{\code{prob}}{The probability of a sample being less than `a`.}
}
}
\description{
plot a normal curve and calculate the probability of a sample being less than
the given value `a`
}
\examples{
myncurve(10, 4, 6)
}
