% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{numeric: the total number of seats available}

\item{gamma}{numeric: the acceptable probability of overbooking}

\item{p}{numeric: the probability that someone who buys a ticket will actually show up}
}
\value{
A named list with the following components:
\describe{
  \item{nd}{numeric; the optimal number of tickets determined using the discrete distribution.}
  \item{nc}{numeric; the optimal number of tickets determined using the normal approximation.}
  \item{N}{numeric; the number of seats available.}
  \item{p}{numeric; the probability of a ticket buyer showing up.}
  \item{gamma}{numeric; the acceptable probability of overbooking.}
}
}
\description{
Determine the number of tickets to sell with a specified success rate, total
number of seats, and overbooking chance tolerance.
}
\examples{
ntickets(N=100, gamma=0.1, p=0.9)
}
