% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmSummaryStats.R
\name{lmSummaryStats}
\alias{lmSummaryStats}
\title{Linear Model Summary Stats}
\usage{
lmSummaryStats(lm)
}
\arguments{
\item{lm}{The linear model to analyze.}
}
\value{
a dataframe with
\itemize{
 \item \code{TSS}
 \item \code{MSS}
 \item \code{RSS}
}
}
\description{
Get the TSS, MSS, and RSS for a linear model.
}
\examples{
df <- data.frame(
  x = rnorm(100, mean = 50, sd = 10),
  y = 5 + 2 * rnorm(100, mean = 50, sd = 10) + rnorm(100, mean = 0, sd = 5)
)

model <- lm(y ~ x, data = df)

lmSummaryStats(model)
}
