% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{mycltb}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{Sample size}

\item{iter}{Number of iterations / samples to take in order to create a histogram of the distribution.}

\item{p}{Probability of success of the binomial distribution}

\item{...}{extra arguments}
}
\value{
w: a vector of the sample means.
}
\description{
Use the central limit theorem to plot the distribution of the mean of samples of the binomial distribution.
}
\examples{
mycltb(n=10, iter=10000, p=0.3)
}
